@@ -2,9 +2,10 @@ name: webdev_example_app
 description: A web app example for webdev CLI.
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dev_dependencies:
-  build_runner: '>=1.6.2 <2.0.0'
-  build_web_compilers: ^2.12.0
-  webdev: ^2.0.0
+  build_runner: ^2.0.0
+  build_web_compilers: ^3.0.0
+  webdev:
+    path: ../webdev
